Add UI improvements, charts, proxy fix and DB seeding
Frontend:
- Professional dark navbar with NavLink active states
- Charts on dashboard (bar + doughnut) using Chart.js
- Confirmation modal when deleting invoices and persons
- Price formatting in tables and detail views
- Live price with VAT preview in invoice form
- Toast notifications replacing alert() and FlashMessage
- Column sorting in invoice and person tables
- Pagination (10 records per page)
- Breadcrumb navigation on detail and form pages
- Invoice filter: InputSelect bug fix, product dropdown
- Proxy fix (pathFilter instead of mounting on /api)

Backend:
- Automatic DB migration on startup
- DatabaseSeeder with test data (5 persons, 10 invoices)

+++ invoice-server-starter/Invoices.Api/Interfaces/IPersonManager.cs
@@ -31,7 +31,7 @@ public interface IPersonManager
 {
     IList<PersonDto> GetAllPersons();
     PersonDto AddPerson(PersonDto personDto);
-    void DeletePerson(uint personId);
+    void DeletePerson(ulong personId);
 
     /// <summary>
     /// Gets a person by ID.
